Libraries (Graphs and Create Functions) closes #8
