+++ Graphs/Class1.cs
@@ -1,0 +1,346 @@
+﻿namespace Graphs
+{
+        public static class GraphicFunc
+        {
+            public static void DrawWelcome()
+            {
+                const string Intro = "Pressiona INTRO per saltar la pantalla en negre", OpenTerminal = "Obre la terminal per a una millor experiencia...", Choose = "Què vols fer?", Play = "Jugar (0)", Leave = "Sortir (1)";
+                const string Screen = $@"
+                             ···························································································
+                             :                                                                                         :
+                             : ██████╗ ███████╗███╗   ██╗██╗   ██╗██╗███╗   ██╗ ██████╗ ██╗   ██╗████████╗     █████╗  :
+                             : ██╔══██╗██╔════╝████╗  ██║██║   ██║██║████╗  ██║██╔════╝ ██║   ██║╚══██╔══╝    ██╔══██╗ :
+                             : ██████╔╝█████╗  ██╔██╗ ██║██║   ██║██║██╔██╗ ██║██║  ███╗██║   ██║   ██║       ███████║ :
+                             : ██╔══██╗██╔══╝  ██║╚██╗██║╚██╗ ██╔╝██║██║╚██╗██║██║   ██║██║   ██║   ██║       ██╔══██║ :
+                             : ██████╔╝███████╗██║ ╚████║ ╚████╔╝ ██║██║ ╚████║╚██████╔╝╚██████╔╝   ██║       ██║  ██║ :
+                             : ╚═════╝ ╚══════╝╚═╝  ╚═══╝  ╚═══╝  ╚═╝╚═╝  ╚═══╝ ╚═════╝  ╚═════╝    ╚═╝       ╚═╝  ╚═╝ :
+                             :                                                                                         :
+                             : ██╗  ██╗███████╗██████╗  ██████╗ ██╗███████╗    ██╗   ██╗███████╗                       :
+                             : ██║  ██║██╔════╝██╔══██╗██╔═══██╗██║██╔════╝    ██║   ██║██╔════╝                       :
+                             : ███████║█████╗  ██████╔╝██║   ██║██║███████╗    ██║   ██║███████╗                       :
+                             : ██╔══██║██╔══╝  ██╔══██╗██║   ██║██║╚════██║    ╚██╗ ██╔╝╚════██║                       :
+                             : ██║  ██║███████╗██║  ██║╚██████╔╝██║███████║     ╚████╔╝ ███████║                       :
+                             : ╚═╝  ╚═╝╚══════╝╚═╝  ╚═╝ ╚═════╝ ╚═╝╚══════╝      ╚═══╝  ╚══════╝                       :
+                             :                                                                                         :
+                             : ███╗   ███╗ ██████╗ ███╗   ██╗███████╗████████╗██████╗ ███████╗███████╗                 :
+                             : ████╗ ████║██╔═══██╗████╗  ██║██╔════╝╚══██╔══╝██╔══██╗██╔════╝██╔════╝                 :
+                             : ██╔████╔██║██║   ██║██╔██╗ ██║███████╗   ██║   ██████╔╝█████╗  ███████╗                 :
+                             : ██║╚██╔╝██║██║   ██║██║╚██╗██║╚════██║   ██║   ██╔══██╗██╔══╝  ╚════██║                 :
+                             : ██║ ╚═╝ ██║╚██████╔╝██║ ╚████║███████║   ██║   ██║  ██║███████╗███████║                 :
+                             : ╚═╝     ╚═╝ ╚═════╝ ╚═╝  ╚═══╝╚══════╝   ╚═╝   ╚═╝  ╚═╝╚══════╝╚══════╝                 :
+                             :                                                                                         :
+                             ···························································································
+            ";
+                const string Pc = @"
+                                                                    _________________
+                                                                   |                 |
+                                                                   |   ___________   |
+                                                                   |  |  Jugar 0  |  |
+                         ______________________________________    |  |___________|  |
+                        |  __________________________________  |   |   ___________   |
+                        | |                                  | |   |  |  Sortir 1 |  |
+                        | |       MONSTRES VS HEROIS         | |   |  |___________|  |
+                        | |              _  _                | |   |   __   __   _   |
+                        | |            (_>()<_)              | |   |  |__| |__| |_|  |
+                        | |             (_/\_)               | |   |                 |
+                        | |               ||                 | |   |                 |
+                        | |             | || |               | |   |                 |
+                        | |            __\||/__              | |   |                 |
+                        | |                                  | |   |       .|.       |
+                        | |                                  | |   |      (   )      |
+                        | |                  By David Sanz   | |   |       '-'       |
+                        | |__________________________________| |   |                 |
+                        |______________________________________|   |                 |
+                                         |    |      '.            |                 |
+                                         |    |        '-.-'-.-'-.-|                 |
+                                         )    (                    |                 |
+                                        /      \                   |                 |
+                                       /________\                  |_________________|
+            ";
+                Console.Clear();
+                Console.ForegroundColor = ConsoleColor.Cyan;
+                Console.WriteLine(Screen);
+                Console.ForegroundColor = ConsoleColor.Green;
+                Console.WriteLine(Intro);
+                Console.ResetColor();
+                Console.ReadLine();
+                Console.WriteLine(OpenTerminal);
+                Console.ReadLine();
+                Console.Clear();
+                Console.ResetColor();
+                Console.WriteLine(Choose);
+                Console.ForegroundColor = ConsoleColor.Green;
+                Console.WriteLine(Play);
+                Console.ForegroundColor = ConsoleColor.Red;
+                Console.WriteLine(Leave);
+                Console.ResetColor();
+                Console.ForegroundColor = ConsoleColor.Gray;
+                Console.Write(Pc);
+                Console.ResetColor();
+            }
+            public static void Archer()
+            {
+                const string Archer = @"          
+                           4$$-.                                          
+                           4   "".                                       
+                           4    ^.                                       
+                           4     $                                       
+                           4     'b                                      
+                           4      ""b.                                   
+                           4        $                                    
+                           4        $r                                   
+                           4        $F                                   
+                -$b========0========$b====*P=->>                         
+                           4       *$$F                                 
+                           4        $$""                                
+                           4       .$F                                  
+                           4       dP                                   
+                           4      F                                     
+                           4     @                                      
+                           4    .                                       
+                           J.  .                                         
+                          '$$
+            ";
+                Console.WriteLine(Archer);
+            }
+            public static void Barbarian()
+            {
+                const string Barbarian = @"
+                                       /\  |\
+                                     __)(__) \
+                                    (__{}__   >
+                                       ||  ) /
+                                       ||  |/
+                                       ||
+                                       ||
+                                       ||
+                                       ||
+                                       ||
+                                       /\
+                                      '--'";
+                Console.WriteLine(Barbarian);
+            }
+            public static void Mage()
+            {
+                const string Mage = @"
+                                        _.-.
+                                    .-'%%~%'.
+                                 _.'%%¨_.'-._\
+                               .'%%%%~|      '
+                              '%¨%%%%%-
+                    .________/_________\_______.
+                     ¨¨°¨°¨¨            ¨°¨¨¨°¨";
+                Console.WriteLine(Mage);
+            }
+            public static void Druid()
+            {
+                const string Druid = @"  
+                                    .   .   .
+                                   / \\/ \\/ \
+                                   \\_     _//
+                                     \\_ _//
+                                       | |
+                                       | |
+                                       | |
+                                       | |
+                                       | |
+                                       | |
+                                       | |
+                                       | |
+                                       | |
+                                       | |
+                                       | |
+                                       | |
+                                       | |
+                                       |_| 
+                                       ( )
+                                      _/ \_
+                                     /_____\ ";
+                Console.WriteLine(Druid);
+            }
+            public static void Monster()
+            {
+                const string Monster = @"
+                                                                  ,--,  ,.-.
+                                   ,                   \,       '-,-`,'-.' | ._
+                                  /|           \    ,   |\         }  )/  / `-,'
+                                  [ ,          |\  /|   | |        /  \|  |/`  ,`
+                                  | |       ,.`  `,` `, | |  _,...(   (      .',
+                                  \  \  __ ,-` `  ,  , `/ |,'      Y     (   /_L\
+                                   \  \_\,``,   ` , ,  /  |         )         _,/
+                                    \  '  `  ,_ _`_,-,<._.<        /         /
+                                     ', `>.,`  `  `   ,., |_      |         /
+                                       \/`  `,   `   ,`  | /__,.-`    _,   `\
+                                   -,-..\  _  \  `  /  ,  / `._) _,-\`       \
+                                    \_,,.) /\    ` /  / ) (-,, ``    ,        |
+                                   ,` )  | \_\       '-`  |  `(               \
+                                  /  /```(   , --, ,' \   |`<`    ,            |
+                                 /  /_,--`\   <\  V /> ,` )<_/)  | \      _____)
+                           ,-, ,`   `   (_,\ \    |   /) / __/  /   `----`
+                          (-, \           ) \ ('_.-._)/ /,`    /
+                          | /  `          `/ \\ V   V, /`     /
+                       ,--\(        ,     <_/`\\     ||      /
+                      (   ,``-     \/|         \-A.A-`|     /
+                     ,>,_ )_,..(    )\          -,,_-`  _--`
+                    (_ \|`   _,/_  /  \_            ,--`
+                     \( `   <.,../`     `-.._   _,-` ";
+                Console.WriteLine(Monster);
+            }
+            public static void StartBattle()
+            {
+                const string StartBattle = @"
+                  ______   ______   .___  ___.  _______ .__   __.   ______  _______ .___  ___.     __          ___         .______        ___   .___________.    ___       __       __          ___       __ 
+                 /      | /  __  \  |   \/   | |   ____||  \ |  |  /      ||   ____||   \/   |    |  |        /   \        |   _  \      /   \  |           |   /   \     |  |     |  |        /   \     |  |
+                |  ,----'|  |  |  | |  \  /  | |  |__   |   \|  | |  ,----'|  |__   |  \  /  |    |  |       /  ^  \       |  |_)  |    /  ^  \ `---|  |----`  /  ^  \    |  |     |  |       /  ^  \    |  |
+                |  |     |  |  |  | |  |\/|  | |   __|  |  . `  | |  |     |   __|  |  |\/|  |    |  |      /  /_\  \      |   _  <    /  /_\  \    |  |      /  /_\  \   |  |     |  |      /  /_\  \   |  |
+                |  `----.|  `--'  | |  |  |  | |  |____ |  |\   | |  `----.|  |____ |  |  |  |    |  `----./  _____  \     |  |_)  |  /  _____  \   |  |     /  _____  \  |  `----.|  `----./  _____  \  |__|
+                 \______| \______/  |__|  |__| |_______||__| \__|  \______||_______||__|  |__|    |_______/__/     \__\    |______/  /__/     \__\  |__|    /__/     \__\ |_______||_______/__/     \__\ (__)
+                ";
+                Console.WriteLine(StartBattle);
+            }
+            public static void Skull()
+            {
+                long skullCount = 5, timerSkull = 50000000;
+                while (skullCount > 0)
+                {
+                    Console.WriteLine(@"
+                                                            ,--.
+                                                           {    }
+                                                           K,   }
+                                                          /  ~Y`
+                                                     ,   /   /
+                                                    {_'-K.__/
+                                                      `/-.__L._
+                                                      /  ' /`\_}
+                                                     /  ' /
+                                             ____   /  ' /
+                                      ,-'~~~~    ~~/  ' /_
+                                    ,'             ``~~~  ',
+                                   (                        Y
+                                  {                         I
+                                 {      -                    `,
+                                 |       ',                   )
+                                 |        |   ,..__      __. Y
+                                 |    .,_./  Y ' / ^Y   J   )|
+                                 \           |' /   |   |   ||
+                                  \          L_/    . _ (_,.'(
+                                   \,   ,      ^^""' / |      )
+                                     \_  \          /,L]     /
+                                       '-_~-,       ` `   ./`
+                                          `'{_            )
+                                              ^^\..___,.--`
+
+                            ");
+                    while (timerSkull > 0)
+                    {
+                        timerSkull--;
+                    }
+                    Console.Clear();
+                    timerSkull = 50000000;
+                    Console.WriteLine(@"
+                                                                        ,--.
+                                                                       {    }
+                                                                       K,   }
+                                                                      /  ~Y`
+                                                                 ,   /   /
+                                                                {_'-K.__/
+                                                                  `/-.__L._
+                                                                  /  ' /`\_}
+                                                                 /  ' /
+                                                         ____   /  ' /
+                                                  ,-'~~~~    ~~/  ' /_
+                                                ,'             ``~~~  ',
+                                               (                        Y
+                                              {                         I
+                                             {      -                    `,
+                                             |       ',                   )
+                                             |        |   ,..__      __. Y
+                                             |    .,_./  Y ' / ^Y   J   )|
+                                             \           |' /   |   |   ||
+                                              \          L_/    . _ (_,.'(
+                                               \,   ,      ^^""' / |      )
+                                                 \_  \          /,L]     /
+                                                   '-_~-,       ` `   ./`
+                                                      `'{_            )
+                                                          ^^\..___,.--`
+
+                            ");
+                    while (timerSkull > 0)
+                    {
+                        timerSkull--;
+                    }
+                    timerSkull = 50000000;
+                    Console.Clear();
+                    skullCount--;
+                }
+                skullCount = 5;
+            }
+            public static void Lose()
+            {
+                const string Lose = @"
+
+
+                             .+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+. 
+                            (                                                                                                             )
+                             )                                                                                                           ( 
+                            (                                                                                                             )
+                             )      ██░ ██  ▄▄▄        ██████     ██▓███  ▓█████  ██▀███  ▓█████▄  █    ██ ▄▄▄█████▓                     ( 
+                            (      ▓██░ ██▒▒████▄    ▒██    ▒    ▓██░  ██▒▓█   ▀ ▓██ ▒ ██▒▒██▀ ██▌ ██  ▓██▒▓  ██▒ ▓▒                      )
+                             )     ▒██▀▀██░▒██  ▀█▄  ░ ▓██▄      ▓██░ ██▓▒▒███   ▓██ ░▄█ ▒░██   █▌▓██  ▒██░▒ ▓██░ ▒░                     ( 
+                            (      ░▓█ ░██ ░██▄▄▄▄██   ▒   ██▒   ▒██▄█▓▒ ▒▒▓█  ▄ ▒██▀▀█▄  ░▓█▄   ▌▓▓█  ░██░░ ▓██▓ ░                       )
+                             )     ░▓█▒░██▓ ▓█   ▓██▒▒██████▒▒   ▒██▒ ░  ░░▒████▒░██▓ ▒██▒░▒████▓ ▒▒█████▓   ▒██▒ ░  ██▓  ██▓  ██▓       ( 
+                            (       ▒ ░░▒░▒ ▒▒   ▓▒█░▒ ▒▓▒ ▒ ░   ▒▓▒░ ░  ░░░ ▒░ ░░ ▒▓ ░▒▓░ ▒▒▓  ▒ ░▒▓▒ ▒ ▒   ▒ ░░    ▒▓▒  ▒▓▒  ▒▓▒        )
+                             )      ▒ ░▒░ ░  ▒   ▒▒ ░░ ░▒  ░ ░   ░▒ ░      ░ ░  ░  ░▒ ░ ▒░ ░ ▒  ▒ ░░▒░ ░ ░     ░     ░▒   ░▒   ░▒        ( 
+                            (       ░  ░░ ░  ░   ▒   ░  ░  ░     ░░          ░     ░░   ░  ░ ░  ░  ░░░ ░ ░   ░       ░    ░    ░          )
+                             )      ░  ░  ░      ░  ░      ░                 ░  ░   ░        ░       ░                ░    ░    ░        ( 
+                            (                                                              ░                          ░    ░    ░         )
+                             )                                                                                                           ( 
+                            (                                                                                                             )
+                             )                                                                                                           ( 
+                            (                                                                                                             )
+                             ""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+""+.+"" 
+
+                    ";
+                Console.WriteLine(Lose);
+            }
+            public static void Win()
+            {
+                const string Congrats = @"
+
+                        ┌──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────┐
+                        │                                                                                                                                                                          │
+                        │  ███████╗███████╗██╗     ██╗ ██████╗██╗████████╗ █████╗ ████████╗███████╗    ██╗  ██╗ █████╗ ███████╗     ██████╗ ██╗   ██╗ █████╗ ███╗   ██╗██╗   ██╗ █████╗ ████████╗  │
+                        │  ██╔════╝██╔════╝██║     ██║██╔════╝██║╚══██╔══╝██╔══██╗╚══██╔══╝██╔════╝    ██║  ██║██╔══██╗██╔════╝    ██╔════╝ ██║   ██║██╔══██╗████╗  ██║╚██╗ ██╔╝██╔══██╗╚══██╔══╝  │
+                        │  █████╗  █████╗  ██║     ██║██║     ██║   ██║   ███████║   ██║   ███████╗    ███████║███████║███████╗    ██║  ███╗██║   ██║███████║██╔██╗ ██║ ╚████╔╝ ███████║   ██║     │
+                        │  ██╔══╝  ██╔══╝  ██║     ██║██║     ██║   ██║   ██╔══██║   ██║   ╚════██║    ██╔══██║██╔══██║╚════██║    ██║   ██║██║   ██║██╔══██║██║╚██╗██║  ╚██╔╝  ██╔══██║   ██║     │
+                        │  ██║     ███████╗███████╗██║╚██████╗██║   ██║   ██║  ██║   ██║   ███████║    ██║  ██║██║  ██║███████║    ╚██████╔╝╚██████╔╝██║  ██║██║ ╚████║   ██║   ██║  ██║   ██║     │
+                        │  ╚═╝     ╚══════╝╚══════╝╚═╝ ╚═════╝╚═╝   ╚═╝   ╚═╝  ╚═╝   ╚═╝   ╚══════╝    ╚═╝  ╚═╝╚═╝  ╚═╝╚══════╝     ╚═════╝  ╚═════╝ ╚═╝  ╚═╝╚═╝  ╚═══╝   ╚═╝   ╚═╝  ╚═╝   ╚═╝     │
+                        │                                                                                                                                                                          │
+                        └──────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────┘
+                    ";
+                const string Medal = @"    
+                                                                                                             _______________
+                                                                                                            |@@@@|     |####|
+                                                                                                            |@@@@|     |####|
+                                                                                                            |@@@@|     |####|
+                                                                                                            \@@@@| ITB |####/
+                                                                                                             \@@@|     |###/
+                                                                                                              `@@|_____|##'
+                                                                                                                   (O)
+                                                                                                                .-'''''-.
+                                                                                                              .'  * * *  `.
+                                                                                                             :  *       *  :
+                                                                                                            : ~ G U A N Y ~ :
+                                                                                                            : ~ A D O R ! ~ :
+                                                                                                             :  *       *  :
+                                                                                                              `.  * * *  .'
+                                                                                                                `-.....-'
+                    ";
+                Console.WriteLine(Congrats);
+                Console.ResetColor();
+                Console.WriteLine();
+                Console.ForegroundColor = ConsoleColor.Yellow;
+                Console.WriteLine(Medal);
+            }
+        }
+}
