+++ ModeFunctions/ModeFunctions.csproj
@@ -1,16 +1,13 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>net7.0</TargetFramework>
-    <RootNamespace>Proyecto_Videojuego</RootNamespace>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <ItemGroup>
     <ProjectReference Include="..\Graphs\Graphs.csproj" />
-    <ProjectReference Include="..\ModeFunctions\ModeFunctions.csproj" />
   </ItemGroup>
 
 </Project>
